Rename sparse window package and classes to slice
