+++ ndarray/src/main/java/org/tensorflow/ndarray/impl/sparse/slice/BooleanSparseSlice.java
@@ -12,34 +12,34 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 =======================================================================*/
-package org.tensorflow.ndarray.impl.sparse.window;
+package org.tensorflow.ndarray.impl.sparse.slice;
 
-import org.tensorflow.ndarray.FloatNdArray;
+import org.tensorflow.ndarray.BooleanNdArray;
 import org.tensorflow.ndarray.NdArray;
 import org.tensorflow.ndarray.NdArrays;
+import org.tensorflow.ndarray.buffer.BooleanDataBuffer;
 import org.tensorflow.ndarray.buffer.DataBuffer;
 import org.tensorflow.ndarray.buffer.DataBuffers;
-import org.tensorflow.ndarray.buffer.FloatDataBuffer;
 import org.tensorflow.ndarray.impl.dimension.DimensionalSpace;
 import org.tensorflow.ndarray.impl.dimension.RelativeDimensionalSpace;
 import org.tensorflow.ndarray.impl.sparse.AbstractSparseNdArray;
 import org.tensorflow.ndarray.index.Index;
 
 import java.nio.ReadOnlyBufferException;
-import java.util.Arrays;
 import java.util.concurrent.atomic.AtomicInteger;
 
-public class FloatSparseWindow extends SparseWindow<Float, FloatNdArray> implements FloatNdArray {
+public class BooleanSparseSlice extends SparseSlice<Boolean, BooleanNdArray>
+    implements BooleanNdArray {
 
   /**
-   * Creates a FloatSparseWindow
+   * Creates a BooleanSparseSlice
    *
-   * @param source the source Sparse Array that this object windows.
+   * @param source the source Sparse Array that this object slices.
    * @param sourcePosition the relative source position into the source
    * @param dimensions the dimensional space for the window
    */
-  public FloatSparseWindow(
-      AbstractSparseNdArray<Float, FloatNdArray> source,
+  public BooleanSparseSlice(
+      AbstractSparseNdArray<Boolean, BooleanNdArray> source,
       long sourcePosition,
       DimensionalSpace dimensions) {
     super(source, sourcePosition, dimensions);
@@ -47,38 +47,37 @@ public class FloatSparseWindow extends SparseWindow<Float, FloatNdArray> impleme
 
   /** {@inheritDoc} */
   @Override
-  public FloatNdArray toDense() {
-    FloatDataBuffer dataBuffer = DataBuffers.ofFloats(shape().size());
+  public BooleanNdArray toDense() {
+    BooleanDataBuffer dataBuffer = DataBuffers.ofBooleans(shape().size());
     read(dataBuffer);
     return NdArrays.wrap(shape(), dataBuffer);
   }
 
   @Override
-  public float getFloat(long... coordinates) {
+  public boolean getBoolean(long... coordinates) {
     return getObject(coordinates);
   }
 
   @Override
-  public FloatNdArray setFloat(float value, long... coordinates) {
+  public BooleanNdArray setBoolean(boolean value, long... coordinates) {
     throw new ReadOnlyBufferException();
   }
 
   @Override
-  public FloatNdArray setObject(Float value, long... coordinates) {
+  public BooleanNdArray setObject(Boolean value, long... coordinates) {
     throw new ReadOnlyBufferException();
   }
 
   @Override
-  public FloatNdArray set(NdArray<Float> src, long... coordinates) {
+  public BooleanNdArray set(NdArray<Boolean> src, long... coordinates) {
     throw new ReadOnlyBufferException();
   }
 
   /** {@inheritDoc} */
   @Override
-  public FloatNdArray read(DataBuffer<Float> dst) {
+  public BooleanNdArray read(DataBuffer<Boolean> dst) {
     // zero out buf.
-    Float[] zeros = new Float[(int) shape().size()];
-    Arrays.fill(zeros, 0f);
+    Boolean[] zeros = new Boolean[(int) shape().size()];
     dst.write(zeros);
 
     AtomicInteger i = new AtomicInteger();
@@ -87,29 +86,29 @@ public class FloatSparseWindow extends SparseWindow<Float, FloatNdArray> impleme
         .forEachIndexed(
             (idx, l) -> {
               long[] coordinates = getIndicesCoordinates(l);
-              float value = getValues().getFloat(i.getAndIncrement());
+              boolean value = getValues().getBoolean(i.getAndIncrement());
               dst.setObject(value, dimensions.positionOf(coordinates));
             });
     return this;
   }
 
   @Override
-  public FloatNdArray read(FloatDataBuffer dst) {
-    return read((DataBuffer<Float>) dst);
+  public BooleanNdArray read(BooleanDataBuffer dst) {
+    return read((DataBuffer<Boolean>) dst);
   }
 
   @Override
-  public FloatNdArray write(DataBuffer<Float> src) {
+  public BooleanNdArray write(DataBuffer<Boolean> src) {
     throw new ReadOnlyBufferException();
   }
 
   @Override
-  public FloatNdArray write(FloatDataBuffer src) {
+  public BooleanNdArray write(BooleanDataBuffer src) {
     throw new ReadOnlyBufferException();
   }
 
   @Override
-  public FloatNdArray slice(Index... indices) {
+  public BooleanNdArray slice(Index... indices) {
     if (indices == null) {
       throw new IllegalArgumentException("Slicing requires at least one index");
     }
@@ -119,18 +118,18 @@ public class FloatSparseWindow extends SparseWindow<Float, FloatNdArray> impleme
 
   /** {@inheritDoc} */
   @Override
-  public FloatNdArray slice(long position, DimensionalSpace sliceDimensions) {
-    return new FloatSparseWindow(this.source, position + sourcePosition, sliceDimensions);
+  public BooleanNdArray slice(long position, DimensionalSpace sliceDimensions) {
+    return new BooleanSparseSlice(this.source, position + sourcePosition, sliceDimensions);
   }
 
   @Override
-  public FloatNdArray get(long... coordinates) {
-    return (FloatNdArray) super.get(coordinates);
+  public BooleanNdArray get(long... coordinates) {
+    return (BooleanNdArray) super.get(coordinates);
   }
 
   @Override
-  public FloatNdArray copyTo(NdArray<Float> dst) {
-    return (FloatNdArray) super.copyTo(dst);
+  public BooleanNdArray copyTo(NdArray<Boolean> dst) {
+    return (BooleanNdArray) super.copyTo(dst);
   }
 
   @Override
